+++ WeatherApp/.vs/WeatherApp/FileContentIndex/0d1995f8-c688-485b-b798-e4f54fcf6ec8.vsidx
