+++ WeatherApp/weatherapp.client/src/App.css
@@ -14,4 +14,3 @@ tr:nth-child(odd) {
     background: #0A0A0A;
 }
 */
-}
