+++ WeatherApp/weatherapp.client/src/index.css
@@ -1,32 +1,26 @@
 :root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  font-weight: 400;
-
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
-
-  font-synthesis: none;
-  text-rendering: optimizeLegibility;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-
+    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+    line-height: 1.5;
+    font-weight: 400;
+    color-scheme: light dark;
+    color: rgba(255, 255, 255, 0.87);
+    background-color: #242424;
+    font-synthesis: none;
+    text-rendering: optimizeLegibility;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
 }
-
-/*a {
-  font-weight: 500;
-  color: #646cff;
-  text-decoration: inherit;
-}
-a:hover {
-  color: #535bf2;
-}*/
 
 /** {
     border: 1px dashed red;
 }
 */
+#MainContents {
+    display: grid;
+    grid-template-columns: minmax(0, 1fr) 350px;
+    grid-column-gap: 12px;
+    max-width: 1162px;
+}
 
 #PageTitle {
     padding: 0px;
@@ -51,32 +45,40 @@ li.NavButton:hover{
     background-color: #808080;
 }
 
-#Weather {
-    display: gird;
-    /*box-sizing: content-box;*/
-    /*width: 60vw;*/
-    /*flex-direction: column;*/
-    /*background-color: gray;*/
+#GridWarp{
+}
+
+#WX_Report{
+    grid-column-start: 1;
+    grid-column-end: 3;
+}
+
+#WeatherCard{
+    /*box-sizing: border-box;*/
+    margin: 12px auto !important;
+    /*background-size: cover;
+    grid-column: span 12;*/
+    grid-column-start: 1;
+    grid-column-end: 3;
 }
 
 #CurrentTemp{
-    /*margin: 0px;
-    padding: 0px;*/
     padding-bottom: 0px;
     margin-bottom: 0px;
     font-size: 80px;
+    text-align: left;
 }
 
 #Temp_Emoji {
-    display: inline-flex;
+    display: flex;
+    justify-content: space-between;
     padding-top: 0px;
     margin-top: 0px;
-    flex-direction: row;
-    align-content: space-between;
-    gap: 60vw;
 }
 
 #WeatherEmoji {
+    display: flex;
+    align-items: center;
     font-size: 70px;
     align-self: flex-end;
     padding-bottom: 0px;
@@ -93,54 +95,3 @@ li.NavButton:hover{
     margin: 0px;
 }
 
-/*th, td {
-    padding-left: 1rem;
-    padding-right: 1rem;
-}
-*/    
-/*body {
-  margin: 0;
-  display: flex;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-}
-*/
-    /*
-  h1 {
-  font-size: 3.2em;
-  line-height: 1.1;
-}
-
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: #1a1a1a;
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
-}
-*/
